refactoringv1
